+++ qoqo-qryd/src/tweezer_devices.rs
@@ -10,6 +10,8 @@
 // express or implied. See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::collections::HashSet;
+
 use bincode::{deserialize, serialize};
 use pyo3::{
     exceptions::{PyTypeError, PyValueError},
@@ -478,6 +480,8 @@ impl TweezerDeviceWrapper {
 
     /// Return the json representation of the TweezerDevice.
     ///
+    /// Additionally, a gate set check is performed.
+    ///
     /// Returns:
     ///     str: The serialized form of TweezerDevice.
     ///
@@ -485,31 +489,22 @@ impl TweezerDeviceWrapper {
     ///     ValueError: Cannot serialize TweezerDevice to json or
     ///         the device does not have valid QRyd gates available.
     fn to_json(&self) -> PyResult<String> {
-        if self.internal.layout_register.values().any(|info| {
-            info.tweezer_single_qubit_gate_times
-                .iter()
-                .all(|(gate_name, _)| {
-                    ALLOWED_NATIVE_SINGLE_QUBIT_GATES.contains(&gate_name.as_str())
-                })
-                && info
-                    .tweezer_two_qubit_gate_times
-                    .iter()
-                    .all(|(gate_name, _)| {
-                        ALLOWED_NATIVE_TWO_QUBIT_GATES.contains(&gate_name.as_str())
-                    })
-                && info
-                    .tweezer_three_qubit_gate_times
-                    .iter()
-                    .all(|(gate_name, _)| {
-                        ALLOWED_NATIVE_THREE_QUBIT_GATES.contains(&gate_name.as_str())
-                    })
-        }) {
-            let serialized = serde_json::to_string(&self.internal).map_err(|_| {
-                PyValueError::new_err("Cannot serialize TweezerMutableDevice to json")
-            })?;
-            Ok(serialized)
-        } else {
-            Err(PyValueError::new_err(
+        let mut all_gates_names: HashSet<&str> = HashSet::new();
+        for layout in self.internal.available_layouts() {
+            all_gates_names.extend(
+                &self
+                    .internal
+                    .get_available_gates_names(Some(layout.to_string()))
+                    .unwrap(),
+            );
+        }
+        if all_gates_names.iter().any(|name| {
+            !ALLOWED_NATIVE_SINGLE_QUBIT_GATES.contains(name)
+                && !ALLOWED_NATIVE_TWO_QUBIT_GATES.contains(name)
+                && !ALLOWED_NATIVE_THREE_QUBIT_GATES.contains(name)
+        }) || all_gates_names.is_empty()
+        {
+            return Err(PyValueError::new_err(
                 "The device does not support valid gates in a layout. ".to_owned()
                     + "The valid gates are: "
                     + &ALLOWED_NATIVE_SINGLE_QUBIT_GATES.join(", ")
@@ -518,11 +513,17 @@ impl TweezerDeviceWrapper {
                     + ", "
                     + &ALLOWED_NATIVE_THREE_QUBIT_GATES.join(", ")
                     + ".",
-            ))
+            ));
         }
+        let serialized = serde_json::to_string(&self.internal)
+            .map_err(|_| PyValueError::new_err("Cannot serialize TweezerMutableDevice to json"))?;
+        Ok(serialized)
     }
 
     /// Convert the json representation of a TweezerDevice to a TweezerDevice.
+    ///
+    /// If a default_layout is found in the input, a layout switch is executed.
+    /// Additionally, a gate set check is performed.
     ///
     /// Args:
     ///     input (str): The serialized TweezerDevice in json form.
@@ -531,12 +532,38 @@ impl TweezerDeviceWrapper {
     ///     TweezerDevice: The deserialized TweezerDevice.
     ///
     /// Raises:
-    ///     ValueError: Input cannot be deserialized to TweezerDevice.
+    ///     ValueError: Input cannot be deserialized to TweezerDevice  or
+    ///         the device does not have valid QRyd gates available.
     #[staticmethod]
     #[pyo3(text_signature = "(input, /)")]
     fn from_json(input: &str) -> PyResult<TweezerDeviceWrapper> {
         let mut internal: TweezerDevice = serde_json::from_str(input)
             .map_err(|_| PyValueError::new_err("Input cannot be deserialized to TweezerDevice"))?;
+        let mut all_gates_names: HashSet<&str> = HashSet::new();
+        for layout in internal.available_layouts() {
+            all_gates_names.extend(
+                &internal
+                    .get_available_gates_names(Some(layout.to_string()))
+                    .unwrap(),
+            );
+        }
+        if all_gates_names.iter().any(|name| {
+            !ALLOWED_NATIVE_SINGLE_QUBIT_GATES.contains(name)
+                && !ALLOWED_NATIVE_TWO_QUBIT_GATES.contains(name)
+                && !ALLOWED_NATIVE_THREE_QUBIT_GATES.contains(name)
+        }) || all_gates_names.is_empty()
+        {
+            return Err(PyValueError::new_err(
+                "The device does not support valid gates in a layout. ".to_owned()
+                    + "The valid gates are: "
+                    + &ALLOWED_NATIVE_SINGLE_QUBIT_GATES.join(", ")
+                    + ", "
+                    + &ALLOWED_NATIVE_TWO_QUBIT_GATES.join(", ")
+                    + ", "
+                    + &ALLOWED_NATIVE_THREE_QUBIT_GATES.join(", ")
+                    + ".",
+            ));
+        }
         if let Some(layout) = &internal.default_layout {
             let _ = internal
                 .switch_layout(&layout.to_string(), None)
@@ -1040,6 +1067,8 @@ impl TweezerMutableDeviceWrapper {
 
     /// Return the json representation of the TweezerMutableDevice.
     ///
+    /// Additionally, a gate set check is performed.
+    ///
     /// Returns:
     ///     str: The serialized form of TweezerMutableDevice.
     ///
@@ -1047,31 +1076,22 @@ impl TweezerMutableDeviceWrapper {
     ///     ValueError: Cannot serialize TweezerMutableDevice to json or
     ///         the device does not have valid QRyd gates available.
     fn to_json(&self) -> PyResult<String> {
-        if self.internal.layout_register.values().any(|info| {
-            info.tweezer_single_qubit_gate_times
-                .iter()
-                .all(|(gate_name, _)| {
-                    ALLOWED_NATIVE_SINGLE_QUBIT_GATES.contains(&gate_name.as_str())
-                })
-                && info
-                    .tweezer_two_qubit_gate_times
-                    .iter()
-                    .all(|(gate_name, _)| {
-                        ALLOWED_NATIVE_TWO_QUBIT_GATES.contains(&gate_name.as_str())
-                    })
-                && info
-                    .tweezer_three_qubit_gate_times
-                    .iter()
-                    .all(|(gate_name, _)| {
-                        ALLOWED_NATIVE_THREE_QUBIT_GATES.contains(&gate_name.as_str())
-                    })
-        }) {
-            let serialized = serde_json::to_string(&self.internal).map_err(|_| {
-                PyValueError::new_err("Cannot serialize TweezerMutableDevice to json")
-            })?;
-            Ok(serialized)
-        } else {
-            Err(PyValueError::new_err(
+        let mut all_gates_names: HashSet<&str> = HashSet::new();
+        for layout in self.internal.available_layouts() {
+            all_gates_names.extend(
+                &self
+                    .internal
+                    .get_available_gates_names(Some(layout.to_string()))
+                    .unwrap(),
+            );
+        }
+        if all_gates_names.iter().any(|name| {
+            !ALLOWED_NATIVE_SINGLE_QUBIT_GATES.contains(name)
+                && !ALLOWED_NATIVE_TWO_QUBIT_GATES.contains(name)
+                && !ALLOWED_NATIVE_THREE_QUBIT_GATES.contains(name)
+        }) || all_gates_names.is_empty()
+        {
+            return Err(PyValueError::new_err(
                 "The device does not support valid gates in a layout. ".to_owned()
                     + "The valid gates are: "
                     + &ALLOWED_NATIVE_SINGLE_QUBIT_GATES.join(", ")
@@ -1080,11 +1100,16 @@ impl TweezerMutableDeviceWrapper {
                     + ", "
                     + &ALLOWED_NATIVE_THREE_QUBIT_GATES.join(", ")
                     + ".",
-            ))
+            ));
         }
+        let serialized = serde_json::to_string(&self.internal)
+            .map_err(|_| PyValueError::new_err("Cannot serialize TweezerMutableDevice to json"))?;
+        Ok(serialized)
     }
 
     /// Convert the json representation of a TweezerMutableDevice to an TweezerMutableDevice.
+    ///
+    /// Additionally, a gate set check is performed.
     ///
     /// Args:
     ///     input (str): The serialized TweezerMutableDevice in json form.
@@ -1093,15 +1118,40 @@ impl TweezerMutableDeviceWrapper {
     ///     TweezerMutableDevice: The deserialized TweezerMutableDevice.
     ///
     /// Raises:
-    ///     ValueError: Input cannot be deserialized to TweezerMutableDevice.
+    ///     ValueError: Input cannot be deserialized to TweezerMutableDevice or
+    ///         the device does not have valid QRyd gates available.
     #[staticmethod]
     #[pyo3(text_signature = "(input, /)")]
     fn from_json(input: &str) -> PyResult<TweezerMutableDeviceWrapper> {
-        Ok(TweezerMutableDeviceWrapper {
-            internal: serde_json::from_str(input).map_err(|_| {
-                PyValueError::new_err("Input cannot be deserialized to TweezerMutableDevice")
-            })?,
-        })
+        let internal: TweezerDevice = serde_json::from_str(input).map_err(|_| {
+            PyValueError::new_err("Input cannot be deserialized to TweezerMutableDevice")
+        })?;
+        let mut all_gates_names: HashSet<&str> = HashSet::new();
+        for layout in internal.available_layouts() {
+            all_gates_names.extend(
+                &internal
+                    .get_available_gates_names(Some(layout.to_string()))
+                    .unwrap(),
+            );
+        }
+        if all_gates_names.iter().any(|name| {
+            !ALLOWED_NATIVE_SINGLE_QUBIT_GATES.contains(name)
+                && !ALLOWED_NATIVE_TWO_QUBIT_GATES.contains(name)
+                && !ALLOWED_NATIVE_THREE_QUBIT_GATES.contains(name)
+        }) || all_gates_names.is_empty()
+        {
+            return Err(PyValueError::new_err(
+                "The device does not support valid gates in a layout. ".to_owned()
+                    + "The valid gates are: "
+                    + &ALLOWED_NATIVE_SINGLE_QUBIT_GATES.join(", ")
+                    + ", "
+                    + &ALLOWED_NATIVE_TWO_QUBIT_GATES.join(", ")
+                    + ", "
+                    + &ALLOWED_NATIVE_THREE_QUBIT_GATES.join(", ")
+                    + ".",
+            ));
+        }
+        Ok(TweezerMutableDeviceWrapper { internal })
     }
 
     /// Return number of qubits in device.